Fix: Pre-rendered index.js page should not cause error
When we copy the pre-rendered pages to `out/`, we do a check to make
sure that the file does not already exist. This is to make sure that we
don't silently overwrite one of the files copied from `public/`, for
example.

But for `index.js` files, Netlify always creates two page entries, one
routing to `/` and one routing to `/index`. So in this case, the first
copying of `index.js` to `out/` goes fine. The second copying fails
because `index.js` already exists.

We fix this by first filtering our list of pages for unique pages and
then performing the copy operation.

Fixes: https://github.com/FinnWoelm/next-on-netlify/issues/2#issuecomment-636415494