Refactor next-on-netlify: Break out into functions
Refactor the next-on-netlify.js file into several functions that are 
easier to manage.

next-on-netlify now creates one Netlify Function per SSR page and API 
endpoint. This is probably better than bundling all pages into one 
function.

Custom _redirects files are no longer merged into the next-on-netlify 
redirects. This is noted in the README and will be re-implemented, if 
demand arises.