Fix redirects for data routes of catch-all pages
When the user navigates client-side (using NextJS Link) to a page that
uses `getServerSideProps` or `getStaticProps`, NextJS makes a request
for the page's data.
This request goes to a special path that follows this pattern:
@_next/data/%BUILD_ID%/path/to/page.json

For pages with fallback: true, we need to redirect that route to the
page's Netlify Function. Dynamic parameters in the data route are
already converted correctly, but until now, this was not true for
catch-all routes.

This commit fixes that by converting data routes with catch-all into
the Netlify-compatible redirect using the * symbol.

For example, /_next/data/%BUILD_ID%/path/to/page/:...slug.json
becomes      /_next/data/%BUILD_ID%/path/to/page/*