Server-side-render pages that use incremental static regeneration
Ideally, pages with incremental static regeneration (ISR) should be
statically pre-rendered and then re-rendered upon request at the
provided revalidation interval. This is currently not possible, because
we can only either have a static page or have a page that is server-side
rendered.

As a temporary workaround, we will be SSRing pages with ISR. This means
they will be less performant that regular static pages, but at least
their content will be fresh. We believe that a focus on the freshness of
the content matches the intention of the user more closely than a focus
on the page's performance.

Discussion: https://github.com/netlify/next-on-netlify/issues/35

Read more about ISR:
https://nextjs.org/blog/next-9-5#stable-incremental-static-regeneration