Expose event + context of Netlify Functions on the req object
When a page is being SSR-ed by a Netlify Function, allow users to access
the function's event and context parameters. These can be accessed as
a property on the `req` object in all SSR-ed pages and in API routes:
- req.netlifyFunction.event
- req.netlifyFunction.context

This allows users to access/leverage Netlify identity for their Next.js
page.
See: https://github.com/netlify/next-on-netlify/issues/20

It also allows users to modify the callbackWaitsForEmptyEventLoop
behavior.
See: https://github.com/netlify/next-on-netlify/issues/66#issuecomment-719988804)